% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_samples.R
\name{annotate_errors}
\alias{annotate_errors}
\title{Annotated F0 data via threshold method}
\usage{
annotate_errors(
  data,
  .unique_file = "uniqueID",
  .hz = "hz",
  .time = "timepoint",
  .samplerate = NA,
  .add_semitones = NA,
  .speaker = NULL,
  rise_threshold = 1.2631578947,
  fall_threshold = 1.7142857143,
  .window_size = 1
)
}
\arguments{
\item{data}{Dataframe}

\item{.unique_file}{Column name containing unique file IDs}

\item{.hz}{Column name containing F0 values in Hz}

\item{.time}{Column name containing timepoint values}

\item{.samplerate}{The sampling rate of the data in milliseconds}

\item{.add_semitones}{If NA, will automatically check to see if \code{F0_semitones}
is present in the dataframe. If it is not, and \code{.speaker} is not NA, the
Hz values will be transformed to semitones using the mean of the speaker.
If \code{.speaker} is NA, the Hz values will be transformed to semitones using
the mean of the unique file ID. If \code{.add_semitones} is TRUE, the Hz values
will always be transformed to semitones, overwriting values in \code{F0_semitones}
if it exists. If \code{.add_semitones} is FALSE, the Hz values will not be
transformed to semitones, even if \code{F0_semitones} is present in the dataframe.}

\item{.speaker}{Column name containing speaker IDs. Set to NA to compute
semitones from the mean of each individual file. Set to NULL to use the
entire dataset to compute the mean.}

\item{rise_threshold}{Rise threshold to use for errors}

\item{fall_threshold}{Fall threshold to use for errors}

\item{.window_size}{Numeric multiplier of the sample rate to ignore pulses.
For example, if you want to ignore pulses that are greater than 1 sample rate
away, leave as 1. If you want to ignore intervals of 80ms and the sampling
rate is 10ms, set to 8.}
}
\value{
Dataframe with annotated F0 data
}
\description{
This function annotates F0 data using the threshold method described in
Steffman and Cole (2022). The implementation here omits diagnostic columns
that are not used in the analysis.
}
\details{
Note that this function is intended to be used with the output of PraatSauce
}
\references{
\itemize{
\item Jeremy Steffman, Jennifer Cole; An automated method for detecting F measurement jumps based on sample-to-sample differences. JASA Express Lett. 1 November 2022; 2 (11): 115201. https://doi.org/10.1121/10.0015045
\item Sundberg, J. (1973). “Data on maximum speed of pitch changes,” Speech Transm. Lab. Q. Prog. Status Rep.4, 39–47.
\item https://github.com/kirbyj/praatsauce
\item https://github.com/jsteffman/f0-jumps
}
}
