% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keybinding_helpers.R
\name{changeKeyBinding}
\alias{changeKeyBinding}
\title{Change key binding}
\usage{
changeKeyBinding(key, bindexpr, msg = NULL)
}
\arguments{
\item{key}{Key to bind}

\item{bindexpr}{Reactive expression to bind to the key}

\item{msg}{Optional message to use}
}
\value{
Nothing, changes boundKeys
}
\description{
In this app we're storing the expressions to execute for each key in an
updateable reactive values object. By doing so we can modify key bindings
more easily later. This function adds a key to the currently tracked keys
and sets the action to execute when the key is pressed.
}
