% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_samples.R
\name{code_carryover_effects}
\alias{code_carryover_effects}
\title{Code carrryover pitch errors}
\usage{
code_carryover_effects(
  data_annotated,
  .unique_file = "uniqueID",
  .hz = "hz",
  .time = "timepoint",
  rise_threshold = 1.2631578947,
  fall_threshold = 1.7142857143
)
}
\arguments{
\item{data_annotated}{Annotated data frame from \code{annotate_errors()}}

\item{.unique_file}{Column name containing unique file IDs}

\item{.hz}{Column name containing F0 values in Hz}

\item{.time}{Column name containing timepoint values}

\item{rise_threshold}{Rise threshold to use for errors}

\item{fall_threshold}{Fall threshold to use for errors}
}
\value{
A data frame with a new column \code{flagged_samples} that codes
potential errors as 1 and non-errors as 0.
}
\description{
This function codes adjacent pitch pulses as a potential error if there was
previously a jump in pitch. For example:
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ Octave jump
           ↓
           OOOO <-the next three should be coded as errors == "carryover errors"
          |    \\
         |      \\
OOOOOOOOO        OOOOOOOOO


Note that potential mistakes arise in cases like below:


O
 \\  All of these might be coded as carryover errors
 \\    ↓↓↓↓
 \\↓↓↓↓OOOO↓↓↓↓
 OOOOO    OOOO↓↓↓↓↓
 ↑            OOOOO
Jump               OOOO <- These might be far enough away (vertically) to
                           stop being coded as potential errors
}\if{html}{\out{</div>}}
}
