% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_samples.R
\name{code_initial_errors}
\alias{code_initial_errors}
\title{Code initial threshold errors}
\usage{
code_initial_errors(
  f0_semitones,
  time,
  time_mutation,
  samplerate,
  windowsize = 8,
  rise_threshold = 1.2631578947,
  fall_threshold = -1.7142857143
)
}
\arguments{
\item{f0_semitones}{Numeric vector}

\item{time}{Numeric vector}

\item{time_mutation}{Mutation factor}

\item{samplerate}{Sampling rate, usually 10}

\item{windowsize}{Buffer size}

\item{rise_threshold}{Positive rise threshold}

\item{fall_threshold}{Negative fall threshold}
}
\value{
Logical vector, TRUE if error
}
\description{
Code initial threshold errors
}
