% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_samples.R
\name{code_threshold_error}
\alias{code_threshold_error}
\title{Code carryover threshold errors}
\usage{
code_threshold_error(
  f0_st,
  f0_err,
  rise_threshold = 1.2631578947,
  fall_threshold = -1.7142857143
)
}
\arguments{
\item{f0_st}{F0 in semitones}

\item{f0_err}{F0 of flagged errors}

\item{rise_threshold}{Positive rise threshold}

\item{fall_threshold}{Negative fall threshold}
}
\value{
Logical vector for threshold errors
}
\description{
Helper to code whether a sample is within the tolerance range for
carryover errors.
}
