% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_module.R
\name{colorServer}
\alias{colorServer}
\title{Color management module server}
\usage{
colorServer(id, plotSettings_ref, updatePlot_reactive, dark_mode_in)
}
\arguments{
\item{id}{module id}

\item{plotSettings_ref}{Reference to the plot settings reactiveValues}

\item{updatePlot_reactive}{A reactive function that updates the plot}

\item{dark_mode_in}{A reactive function that returns the current theme,
should be a reactive wrapper around the dark mode input}
}
\value{
A module server function
}
\description{
This module server takes care of the following functionality:
}
\details{
\itemize{
\item Manage the color pickers on the settings page
\item Set the color pickers and plot settings when the user changes the theme
\item Set the correct color for the load files button
}
}
