% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_color.R
\name{colorUI}
\alias{colorUI}
\title{Color management module UI}
\usage{
colorUI(id)
}
\arguments{
\item{id}{id for namespace}
}
\value{
UI elements for color management
}
\description{
Renders the three color pickers on the settings pane
}
