% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demoEditor.R
\name{demoEditor}
\alias{demoEditor}
\title{Open web-version editor}
\usage{
demoEditor(
  input_directory = get_example_f0_data("."),
  output_directory = "./",
  audio_directory = "./audio",
  textgrid_directory = "./audio",
  praat_path = "./Praat.exe",
  ...
)
}
\arguments{
\item{input_directory}{Directory containing the csv files to load}

\item{output_directory}{Directory to save annotated csv files to}

\item{audio_directory}{Directory containing .wav files to play or send to
praat}

\item{textgrid_directory}{Directory containing .TextGrid files to send to
praat}

\item{praat_path}{For Windows, the path to a Praat executable. For linux and
mac, this should just be Praat assuming it's available as a terminal command.}

\item{...}{not used}
}
\value{
Nothing
}
\description{
Runs the web app version of the app
}
