% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editor_web.R
\name{demoEditor_legacy}
\alias{demoEditor_legacy}
\title{Open web-version editor}
\usage{
demoEditor_legacy(...)
}
\arguments{
\item{...}{not used}
}
\value{
Nothing
}
\description{
Runs the web app version of the app
}
