% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_samples.R
\name{.convert_time_to_ms}
\alias{.convert_time_to_ms}
\title{Convert seconds (probably) to milliseconds}
\usage{
.convert_time_to_ms(
  data,
  .time,
  .samplerate,
  only_samplerate = FALSE,
  .verbose = FALSE
)
}
\arguments{
\item{data}{Dataframe}

\item{.time}{column name to use for time}

\item{.samplerate}{sampling rate for the unit of time in the dataframe,
so for a 10ms sampling rate with times in ms, this should be 10; if it's
recorded in seconds, this should be 0.01}

\item{only_samplerate}{Whether to return only the sampling rate, default FALSE}

\item{.verbose}{Logical, default FALSE, whether to print messages}
}
\value{
A list containing the transformed time and the sampling
rate in milliseconds
}
\description{
The flag samples algorithm assumes that the time is in milliseconds, but it
is not uncommon for time to be specified in seconds. This function converts
the time to milliseconds if it is not already in milliseconds and automatically
tries to detect the sampling rate if it is not specified. The sampling rate
is determined by getting the median of the sample-to-sample time differences,
as it's most likely that the VAST majority of these will be the sampling rate
}
