% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{find_path}
\alias{find_path}
\title{find path}
\usage{
find_path(pitchobj, filename)
}
\arguments{
\item{pitchobj}{A pitch object (i.e., a List) from rPraat::pitch.read or PitchMendR::pitch.read2}

\item{filename}{Character vector of length 1, filename to use}
}
\description{
creates a data.table by loading the relevant f0 info from the pitch obj
}
