% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_samples.R
\name{flag_potential_errors}
\alias{flag_potential_errors}
\title{Flag errors wrapper}
\usage{
flag_potential_errors(
  data,
  .unique_file = "uniqueID",
  .hz = "hz",
  .time = "timepoint",
  .samplerate = NA,
  .add_semitones = NA,
  .speaker = NA,
  rise_threshold = 1.2631578947,
  fall_threshold = 1.7142857143
)
}
\arguments{
\item{data}{Data frame}

\item{.unique_file}{Column name containing unique file IDs}

\item{.hz}{Column name containing F0 in Hz}

\item{.time}{Column name containing timepoint}

\item{.samplerate}{Sample rate in milliseconds (usually 10ms)}

\item{.add_semitones}{Add semitones to F0}

\item{.speaker}{Column name containing Speaker ID}

\item{rise_threshold}{Rise threshold to use for errors}

\item{fall_threshold}{Fall threshold to use for errors}
}
\value{
Dataframe with potential errors coded in the \code{flagged_samples} column
}
\description{
Flags potential sample-to-sample errors in F0 data. Uses the algorithm
described in Steffman & Cole (2022) but uses a faster implementation that
doesn't rely on NAs as much. Note however that while this is faster, it is
written to be as close to the original implementation as possible, so there
are some quirks and inefficiencies that remain.
}
\references{
\itemize{
\item Jeremy Steffman, Jennifer Cole; An automated method for detecting F measurement jumps based on sample-to-sample differences. JASA Express Lett. 1 November 2022; 2 (11): 115201. https://doi.org/10.1121/10.0015045
\item https://github.com/kirbyj/praatsauce
\item https://github.com/jsteffman/f0-jumps
}
}
