% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\name{get_example_f0_data}
\alias{get_example_f0_data}
\title{Load example data path}
\usage{
get_example_f0_data(path = NULL)
}
\arguments{
\item{path}{Path to search for extdata in}
}
\value{
Path to \code{demo_data.csv}
}
\description{
Load example data path
}
