% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_value_octave_away.R
\name{get_value_octave_away}
\alias{get_value_octave_away}
\title{Get pitch candidate an octave away from current F0}
\usage{
get_value_octave_away(frame, octave, threshold = 0.25)
}
\arguments{
\item{frame}{List from pitch object. E.g., \code{pitchObj$frame[[i]]}}

\item{octave}{String, either "halve" or "double"}

\item{threshold}{Threshold value in semitones, default .25 (25 cents) but subject to tuning}
}
\value{
integer index of the candidate an octave away. If no such
candidate is found, returns 1 (index of the current f0 value).
}
\description{
Given a frame containing pitch candidate information and a desired octave
transformation (either "halve" or "double), look up which candidate (if any)
corresponds to that transformation. E.g., if F0 is 110 and we want to double,
this will look up which candidate is equal to 220. Because getting values at
ratios of exactly 2 or 1/2 is unlikely, a small threshold is added. So, this
will actually look for a candidate between \verb{[220-threshold, 220+threshold]}

This shouldn't be called by itself, only within get_and_set_freq_values in
module_sauce_octaveshift.R
}
