% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hz_to_semitones.R
\name{hz_to_semitones}
\alias{hz_to_semitones}
\title{Calculate semitone differences}
\usage{
hz_to_semitones(
  hz_vals,
  .semitones_from = NA,
  .n_starting_samples = 5L,
  .startmethod = stats::median,
  .quiet = FALSE
)
}
\arguments{
\item{hz_vals}{Numeric vector of Hz values}

\item{.semitones_from}{Value to calculate semitones from. If left as NA,
a baseline will be created by using \code{.startmethod} with the first
\code{.n_starting_samples} pitch points}

\item{.n_starting_samples}{Integer number of starting pitch points to use
for calculating the baseline}

\item{.startmethod}{Function to create the baseline. Default is \code{median} with
the first \code{.n_starting_samples}. Other recommendations are \code{mean} or \code{first}.}

\item{.quiet}{Whether to message user about value selected for baseline}
}
\value{
Vector of semitone values
}
\description{
Calculates a vector of Hz values to semitone differences from a baseline.
Resulting vector includes a \code{semitones_from} attribute.
Access with \code{attr(semitone_vector, 'semitones_from')}.
}
\examples{

hz_to_semitones(c(110, 220, 440, 540), .semitones_from = 110)

}
