% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyboard_css.R
\name{kbd}
\alias{kbd}
\title{Insert a kbd button}
\usage{
kbd(key, class = NULL, size = "sm")
}
\arguments{
\item{key}{String denoting a key to use, e.g., 'k' or 'ctrl'. If passed a
character vector, only the first element is used.}

\item{class}{Additional classes to set on the button. Default NULL will
set class kbc-button.}

\item{size}{Size to apply to the button. Default "sm" will set class
kbc-button-sm. See https://github.com/shhdharmen/keyboard-css for more info}
}
\value{
A kbd tag with the appropriate classes and key
}
\description{
This function creates a kbd tag using keyboard-css.
This type of button will not animate when clicked.
}
