% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyboard_css.R
\name{kbd_button}
\alias{kbd_button}
\title{Insert a kbd button}
\usage{
kbd_button(key, class = NULL, size = "sm", id = NULL)
}
\arguments{
\item{key}{String denoting a key to use, e.g., 'k' or 'ctrl'. If passed a
character vector, only the first element is used.}

\item{class}{Additional classes to set on the button. Default NULL will
set class kbc-button.}

\item{size}{Size to apply to the button. Default "sm" will set class
kbc-button-sm. See https://github.com/shhdharmen/keyboard-css for more info}

\item{id}{An id to set on the button. This is useful for adding event
listeners or other JavaScript functionality.}
}
\value{
A button tag with the appropriate classes and key
}
\description{
This function creates a button tag using keyboard-css.
This type of button will animate when clicked.
You can pass an id to add event listeners elsewhere in a server.
}
