% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keybinding_helpers.R
\name{keyBindAction}
\alias{keyBindAction}
\title{Create a key bind action with optional message}
\usage{
keyBindAction(bindexpr, msg = NULL)
}
\arguments{
\item{bindexpr}{Reactive expression to execute}

\item{msg}{Message to use, no message is sent if NULL}
}
\value{
A reactive that also sends a message
}
\description{
Bundles together a message with a reactive expression to be used as a key
binding action.
}
