% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app - Copy.R, R/app.R
\name{openEditor}
\alias{openEditor}
\title{Run pulse editor GUI}
\usage{
openEditor(...)

openEditor(...)
}
\arguments{
\item{...}{not used}
}
\value{
Nothing

Nothing
}
\description{
Runs the shiny app

Runs the shiny app
}
