% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editor_sauce.R
\name{openSauceEditor}
\alias{openSauceEditor}
\title{Run pulse editor GUI}
\usage{
openSauceEditor(
  input_directory = get_example_f0_data("."),
  output_directory = "./",
  audio_directory = "./audio",
  textgrid_directory = "./audio",
  praat_path = "./Praat.exe",
  yvar = "f0",
  xvar = "t",
  ...
)
}
\arguments{
\item{input_directory}{Directory containing the csv files to load}

\item{output_directory}{Directory to save annotated csv files to}

\item{audio_directory}{Directory containing .wav files to play or send to
praat}

\item{textgrid_directory}{Directory containing .TextGrid files to send to
praat}

\item{praat_path}{For Windows, the path to a Praat executable. For linux and
mac, this should just be Praat assuming it's available as a terminal command.}

\item{yvar}{Y variable to use, should exist in the file to be loaded. Default
is \code{'f0'}.}

\item{xvar}{X variable touse, should exist in the file to be loaded. Default
is \code{'t_ms'}.}

\item{...}{not used}
}
\value{
Nothing
}
\description{
Runs the shiny app
}
