% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_samples.R
\name{propagate_f0_of_err}
\alias{propagate_f0_of_err}
\title{Propagate F0 errors rightward}
\usage{
propagate_f0_of_err(F0_of_err)
}
\arguments{
\item{F0_of_err}{Numeric vector, F0_of_err column}
}
\value{
Numeric vector
}
\description{
If the current F0 value is 0, look at the previous value and use that instead.
This proceeds from left to right, propagating errors forward through sequences
of 0s. This replaces the step in the original implementation where this was
calculated through repeated lead/lag transformations.
}
