% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PulseTransformR-package.R
\docType{package}
\name{PulseTransformR-package}
\alias{PulseTransformR}
\alias{PulseTransformR-package}
\title{PulseTransformR: Fix Pitch Tracking Errors}
\description{
This package provides an interface to modify pitch samples saved in a csv file. Users can remove spurious tracking errors and non-destructively fix halving/doubling errors.
}
\author{
\strong{Maintainer}: Thomas Sostarics \email{tsostarics@gmail.com} (\href{https://orcid.org/0000-0002-1178-7967}{ORCID})

}
\keyword{internal}
