% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_praat_script_utils.R
\name{run_temp_script}
\alias{run_temp_script}
\title{Run a praat script that deletes itself}
\usage{
run_temp_script(lines, praat_path = "./Praat.exe")
}
\arguments{
\item{lines}{Character vector, each string is a line of the praat script}

\item{praat_path}{Path to praat, use audioInfo$praatPath() outside of the
praat module}
}
\value{
Invisibly returns 0
}
\description{
Praat streams in the script lines instead of loading the whole file into
memory, so we need to wait until the script is done running before we try and
delete the file. BUT, if we use wait = TRUE, the script won't send back a
return value even if you use exitScript() as the last line. To get around
this, we have the praat script delete itself when it's done executing (hence
the deleteFile line) and then wait until the file stops existing so we can
continue.
}
