% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textSubmitButton.R
\name{submitTextInput}
\alias{submitTextInput}
\title{Text input with submit button}
\usage{
submitTextInput(
  inputId,
  label,
  value = "",
  width = NULL,
  icon = "share-from-square",
  placeholder = "Enter text here",
  disabled = FALSE,
  ...
)
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{label}{Display label for the control, or \code{NULL} for no label.}

\item{value}{Initial value.}

\item{width}{The width of the input, e.g. \code{'400px'}, or \code{'100\%'};
see \code{\link[=validateCssUnit]{validateCssUnit()}}.}

\item{icon}{The icon used for the button. Default share-from-square}

\item{placeholder}{A character string giving the user a hint as to what can
be entered into the control. Internet Explorer 8 and 9 do not support this}

\item{disabled}{If \code{TRUE}, the button will not be clickable. Use
\code{\link[=updateActionButton]{updateActionButton()}} to dynamically enable/disable the button.}

\item{...}{Named attributes to be applied to the button or link.
option.}
}
\description{
Modified from the source for fileInput: https://github.com/rstudio/shiny/blob/main/R/input-file.R
}
\details{
Create a text input with a submit button to the left of the text box.
The button can be styled with a font-awesome icon.
The text value's inputId is the inputId field, the button's inputId has
\code{Button} appended to it. E.g., inputId = myTextInput yields input$myTextInput
and input$myTextInputButton
}
\concept{input elements}
