% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{swapFrameValue}
\alias{swapFrameValue}
\title{Swap frame candidates}
\usage{
swapFrameValue(frame, i, j)
}
\arguments{
\item{frame}{List of frames from a Pitch object}

\item{i}{first index, must be integer}

\item{j}{second index, must be integer}
}
\description{
Given a list of frames, swap the candidates at the given 1-indexed positions
in place and then return the new value at i
}
