% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_cdf.R
\name{update_cdf}
\alias{update_cdf}
\title{Update candidate dataframe}
\usage{
update_cdf(rawPitchDB, plotted_file, which_frame, new_candidate_i)
}
\arguments{
\item{rawPitchDB}{rawPitchDB}

\item{plotted_file}{File to use to index into the database}

\item{which_frame}{which frame to subset to}

\item{new_candidate_i}{index of the new candidate, typically not 1L}
}
\value{
nothing, updates specified candidate dataframe in place
}
\description{
Updates a candidate data frame in the raw pitch database by swapping the
first index with the given candidate
}
