% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyboard_css.R
\name{use_keyboardcss}
\alias{use_keyboardcss}
\title{Use keyboard-css from local installation}
\usage{
use_keyboardcss()
}
\value{
A singleton tagList for the UI
}
\description{
You must place this somewhere in your UI to use keyboard-css.
This will load the keyboard-css stylesheet from the one saved in the package.
}
