% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyboard_css.R
\name{use_keyboardcss_link}
\alias{use_keyboardcss_link}
\title{Use keyboard-css from online build}
\usage{
use_keyboardcss_link()
}
\value{
A singleton tagList for the UI
}
\description{
If you don't want to use the local version for some reason, you can use
this function to load the keyboard-css stylesheet from the online build.
You must place either this or \code{use_keyboardcss} somewhere in your UI.
}
