% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoom_controls.R
\name{zoom_by}
\alias{zoom_by}
\title{Zoom by}
\usage{
zoom_by(current_xmin, current_xmax, factor = 1)
}
\arguments{
\item{current_xmin}{Current xmin of the plot}

\item{current_xmax}{Current xmax of the eplot}

\item{factor}{Factor to zoom in by, default 1 (=no zoom), but
will typically be either 0.5 (zoom in) or 2.0 (zoom out)}
}
\value{
Numeric vector of length 2 with the new x range values to use
}
\description{
Adapted from https://github.com/praat/praat/blob/8ef0fca519d833d55729a339bd73226074215fe7/foned/FunctionEditor.cpp#L471
so the zooming behavior is the same as that in Praat.
}
