% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hz_to_semitones.R
\name{.var_of_diffs}
\alias{.var_of_diffs}
\title{Variance of semitone difference}
\usage{
.var_of_diffs(x)
}
\arguments{
\item{x}{Hz values}
}
\value{
Variance of semitone differences
}
\description{
A heuristic calculation of the variance of the sample-to-sample semitone
differences. Contours with fewer extreme changes in semitones will have
lower variance.
}
