% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_samples.R
\name{flag_potential_errors}
\alias{flag_potential_errors}
\title{Flag errors wrapper}
\usage{
flag_potential_errors(
  data,
  .unique_file = "uniqueID",
  .hz = "hz",
  .time = "timepoint",
  .samplerate = NA,
  .add_semitones = NA,
  .speaker = NA,
  rise_threshold = 1.2631578947,
  fall_threshold = 1.7142857143,
  .as_vec = FALSE,
  .ignore_0s = TRUE,
  .window_size = 8
)
}
\arguments{
\item{data}{Data frame}

\item{.unique_file}{Column name containing unique file IDs}

\item{.hz}{Column name containing F0 in Hz}

\item{.time}{Column name containing timepoint}

\item{.samplerate}{Sample rate in milliseconds (usually 10ms)}

\item{.add_semitones}{Add semitones to F0}

\item{.speaker}{Column name containing Speaker ID}

\item{rise_threshold}{Rise threshold to use for errors}

\item{fall_threshold}{Fall threshold to use for errors}

\item{.as_vec}{Whether just the results should be returned as a logical vector
(TRUE) or if the whole dataframe should be returned (FALSE)}

\item{.ignore_0s}{If TRUE, will separate the dataset into two parts: where
.hz is equal to 0 or NA, and where it is not (i.e., actual F0 values). The
algorithm will run with the latter set, then the former will be merged back
into the final dataset. This way the missing values are ignored, but the
results are of the same size as the input. If this is set to FALSE, then
you will get more false positives due to having many more F0 jumps from a
starting value of 0.}

\item{.window_size}{Numeric multiplier of the sample rate to ignore pulses.
For example, if you want to ignore pulses that are greater than 1 sample rate
away, leave as 1. If you want to ignore intervals of 80ms and the sampling
rate is 10ms, set to 8.}
}
\value{
Dataframe with potential errors coded in the \code{flagged_samples} column
}
\description{
Flags potential sample-to-sample errors in F0 data. Uses a modified version
of the algorithm described in Steffman & Cole (2022). The implementaion here
is much faster than the original version and avoids NA issues and
overprocessing short files when longer files exist.
}
\references{
\itemize{
\item Jeremy Steffman, Jennifer Cole; An automated method for detecting F measurement jumps based on sample-to-sample differences. JASA Express Lett. 1 November 2022; 2 (11): 115201. https://doi.org/10.1121/10.0015045
\item https://github.com/kirbyj/praatsauce
\item https://github.com/jsteffman/f0-jumps
}
}
