% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyboard_css.R
\name{inline_kbd_button}
\alias{inline_kbd_button}
\title{Interpolate string with inline kbd button}
\usage{
inline_kbd_button(key, string, ...)
}
\arguments{
\item{key}{Character vector of keys to display}

\item{string}{A glue string to interpolate. Use \code{{KEY}} to insert the inline
key if you're only using one key, or \code{{KEY[1]}}, \code{{KEY[2]}}, etc. if you're
inserting multiple keys. Note that you are only allowed to interpolate
\code{KEY}, other R code will not be evaluated.}

\item{...}{Additional arguments passed to \code{kbd_button()}}
}
\value{
An interpolated string with inline kbd buttons
}
\description{
Inserts inline kbd buttons
}
\examples{
\dontrun{

inline_kbd_button("a", "Press {KEY} to continue")
# > Press <button class="kbc-button kbc-button-sm" style="width:fit-content">A</button> to continue

inline_kbd_button(c("ctrl", "z"), "Press {KEY[1]}+{KEY[2]} to undo")
# > Press
<button class="kbc-button kbc-button-sm" style="width:fit-content">CTRL</button>
+
<button class="kbc-button kbc-button-sm" style="width:fit-content">Z</button>
to undo
}
}
