% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{openEditor}
\alias{openEditor}
\title{Run pulse editor GUI}
\usage{
openEditor(
  input_dir = get_example_f0_data("."),
  output_dir = "./",
  praat_path = "./Praat.exe",
  audio_directory = "./audio",
  textgrid_directory = "./audio",
  ...
)
}
\arguments{
\item{...}{not used}
}
\value{
Nothing
}
\description{
Runs the shiny app
}
